Refactor the containers module into appropriate modules
