+++ docs/repobee_plug/code.rst
@@ -41,6 +41,12 @@ name
 .. automodule:: repobee_plug.name
     :members:
 
+reviews
+=======
+
+.. automodule:: repobee_plug.reviews
+    :members:
+
 cli
 ===
