+++ src/repobee_plug/_exthooks.py
@@ -12,9 +12,9 @@ import argparse
 import configparser
 from typing import Optional
 
+from repobee_plug.cli.args import ConfigurableArguments
 from repobee_plug.platform import PlatformAPI
-from repobee_plug._containers import hookspec
-from repobee_plug._containers import Result, ConfigurableArguments
+from repobee_plug.hook import hookspec, Result
 from repobee_plug._deprecation import deprecate
 
 from repobee_plug.localreps import StudentRepo, TemplateRepo
