+++ src/repobee_plug/cli/__init__.py
@@ -14,6 +14,7 @@ from repobee_plug.cli.args import (
 from repobee_plug.cli.commandmarkers import Command, CommandExtension
 
 from repobee_plug.cli._corecommand import _CoreCommand
+from repobee_plug.cli._corecommand import BaseParser
 
 CoreCommand = _CoreCommand()
 
@@ -29,5 +30,6 @@ __all__ = [
     "Command",
     "CommandExtension",
     "CoreCommand",
+    "BaseParser",
     "is_cli_arg",
 ]
