+++ src/repobee_plug/cli/args.py
@@ -263,3 +263,11 @@ def mutually_exclusive_group(*, __required__: bool = False, **kwargs):
     """
     options = [(key, value) for key, value in kwargs.items()]
     return MutuallyExclusiveGroup(required=__required__, options=options)
+
+
+@dataclasses.dataclass(frozen=True)
+class ConfigurableArguments:
+    """A container for holding a plugin's configurable arguments."""
+
+    config_section_name: str
+    argnames: List[str]
