+++ tests/unit_tests/repobee_plug/test_serialize.py
@@ -1,8 +1,7 @@
 import pytest
 import collections
 
-from repobee_plug import _containers
-from repobee_plug import _serialize
+import repobee_plug as plug
 
 
 @pytest.fixture
@@ -12,29 +11,27 @@ def hook_result_mapping():
         (
             "slarse-task-1",
             "junit4",
-            _containers.Status.SUCCESS,
+            plug.Status.SUCCESS,
             "All tests passed",
             {"extra": "data", "arbitrary": {"nesting": "here"}},
         ),
         (
             "slarse-task-1",
             "javac",
-            _containers.Status.ERROR,
+            plug.Status.ERROR,
             "Some stuff failed",
             None,
         ),
         (
             "glassey-task-2",
             "pylint",
-            _containers.Status.WARNING,
+            plug.Status.WARNING,
             "-10/10 code quality",
             None,
         ),
     ]:
         hook_results[repo_name].append(
-            _containers.Result(
-                name=hook_name, status=status, msg=msg, data=data
-            )
+            plug.Result(name=hook_name, status=status, msg=msg, data=data)
         )
     return {
         repo_name: sorted(results)
@@ -42,12 +39,12 @@ def hook_result_mapping():
     }
 
 
-def test_serialize_empty_mapping():
-    assert _serialize.result_mapping_to_json({}) == "{}"
+def testplug_empty_mapping():
+    assert plug.result_mapping_to_json({}) == "{}"
 
 
 def test_desezialize_empty_json():
-    assert _serialize.json_to_result_mapping("{}") == {}
+    assert plug.json_to_result_mapping("{}") == {}
 
 
 def test_lossless_serialization(hook_result_mapping):
@@ -56,8 +53,8 @@ def test_lossless_serialization(hook_result_mapping):
     """
     expected = dict(hook_result_mapping)
 
-    serialized = _serialize.result_mapping_to_json(hook_result_mapping)
-    deserialized = _serialize.json_to_result_mapping(serialized)
+    serialized = plug.result_mapping_to_json(hook_result_mapping)
+    deserialized = plug.json_to_result_mapping(serialized)
     actual = {
         repo_name: sorted(results)
         for repo_name, results in deserialized.items()
